changed methods remove and added implementation for the missing exception
as you said when we remove an order, we have to remove transports associated as well.
but not always.
when we remove for example a transport, should we remove the routes associated? why?
each routes could remain avaiable for another transport, i dont think we have to remove it even if is in the transport removed